Add the very first section of the project
The very first section was added to the project.

The very first section of the project contains a header on the top which
includes the logo of the page which is "LoveThunder&Rain", three
hyperlinks to all the different topics and th section itself which
brings a curious fact about lightning with a button and invites the
reader to discover more about thunders, and last but not least, an image
of a lightning.